add change password option into the profile section
